LS top-n annotation (WIP)
