+++ two_stage_pipeliner/data_converters/label_studio.py
@@ -25,7 +25,7 @@ class LabelStudioDetectionDataConverter(DataConverter):
     @DataConverter.assert_image_data
     def get_image_data_from_annot(
         self,
-        image_path: Union[Path, str],
+        image_path: Union[str, Path],
         annot: Union[Path, str, Dict],
     ) -> ImageData:
         if isinstance(annot, str) or isinstance(annot, Path):
